Merge PRs with the opam-repository master branch
It now uses the upstream OCurrent, since the merging is now done by the
CI, not in OCurrent. When `master` updates, the rebuild is done in the
background (it still shows as successful in GitHub until the new result
is known, rather than switching to pending).

The `sed` invocation didn't work for me, so I replaced it with some
OCaml code, which is probably neater anyway. I added a check that the
opam file is version 2 and that it's in the right place.

I updated the instructions for reproducing the build locally.

With these changes, I was able to run opam-repo-ci-local on a local
clone of opam-repository.

Note: this also removes the `~enable_submodules` option. We might want
to bring that back at some point, but I suspect that people trying to
add submodules to opam-repository is not an urgent problem at the
moment. We may want to upstream other changes from the branch at some
point too, but let's get this working first...