Remove nested use of Current.t from pipeline
This removes the `pipeline` type and its nested currents. Instead, the
result is a simple current whose value is the tree of labels, jobs IDs
and results.

It also does the pull only once per package/platform, and adds some
labels to the diagrams. This makes the diagrams quite readable with the
plain upstream OCurrent. Eventually, the base image pulls should be
moved to the top-level, as in ocaml-ci.