Improve local test command
The existing command hasn't worked since the move to OCluster, because
the cluster requires the URL of a public repository, not a path on your
local machine.

This replaces the old `opam-repo-ci-local` command with one that works
more like the main `opam-repo-ci-service` command. In particular, it
tests a whole repository and updates the index, so it works with the web
UI. However, it only requires an OAuth token (not a full app) and it
doesn't push statuses back to GitHub.