web-ui: don't show the tree of other builds on the build page
For opam-repo-ci the list of builds is so long it fills the page, so
that clicking on a particular build doesn't appear to do anything.

With this change, clicking on a build shows just the log for that build.
You can use the breadcrumbs at the top to return to the list of builds.