Latch master commit
We track `master` and merge PRs against it for testing. However,
existing builds are not redone when it changes (because that wastes too
much time).

Unfortunately, when master updates there is a brief period when we know
that a new commit is available but we haven't fetched it yet. During
this time all analysis jobs become blocked, which in turn blocks all
builds. Any builds currently in progress are cancelled. When the fetch
of master completes, the analysis step becomes active again and all jobs
are restarted.

This commit adds an in-memory latch of the master commit, so that we
continue passing the old version of master to analyse until the new one
is ready.