Remove the revdep on successful install
To be able to use the same switch for all the revdeps tests, we remove
the revdep that was installed in the current test, so that we can
install the new revdep in the same switch.

We run a remove with autoremove, so any additional packages that were
installed should also be removed, reducing the chances for conflicts or
unnecessary recompiles of packages that we don't care about in a
particular test.