Don't run on Oracle Linux
Closes tarides/infrastructure#320

This is a followup correcting the intended logic. Code looked correct,
but it was actually failing to filter out Oracle due to the precedence
of `when` clauses and disjunctive cases: `| x | y | z when cond -> foo`
means `(| x | y | z) when cond -> foo | _ -> bar`, rather than
`| x | y | (z when cond) -> foo`. E.g.,:

```
match 1 with
 | 1 | 2 | _ when false -> "foo"
 | _ -> "bar"
;;
- : string = "bar"
```