Refactor Test logic
The aim of this change is to make the API of the `Test` module more
useful to other programs in the future. The main change to effect this
is to have `Test.test_packages_with_opam` return the data from install
failures, rather than just reporting this by printing to stdout.

Because the testing process can take some time for each installation, we
have opted to report this as a sequence, rather than a list, so that
each error can be handled as it it returned, rather than waiting for all
the tests to complete.

A change suggested along the way is to do the repo configuration and
installation of the target package once, rather than repeat it for each
revdep tested.