Add build test command for local runs
This commit exposes the functionality to run a CI build/test locally by
generating an Obuilder spec and dumping it to a local Dockerfile and
building it. We create the Dockerfile (and the dockerignore) in a tmp
directory and use them to build the container, with the opam-repository
directory as the docker context.

Co-authored-by: Shon Feder <shon.feder@gmail.com>