Simplify pkgopt construction
The `~pkgopt` argument required by the `build` function is being
constructed twice, in the exact same way, in the two location where
`build` is applied. Additionally, its only dependency is the `analysis`,
which is already supplied to `build`, and its only depenedent is
`build`. In other words, from the data flow, `pkgopts` is a
subcomputation of the `build` process, and so should be located the body
of the latter.

Additionally, within the build we have a "dummy" dependency on
`analysis` which as commented as being required only to structure the
report in a certain way. However, the derivation of `pkgopts` has a real
dependency on `analysis`, so we don't need the dummy one.