Embedd git hash of opam-ci-check version
When running `opam-ci-check` in workers, we need to ensure we are
running the right version of that tool. We cannot simply pin it to the
`live` branch, because obuilder/docker might cache a previous install
from the command, and fail to pick up an update. On the other hand, we
do want caching, so we don't want to always install fresh.

As a workaround to meet these requirements given our current code
organization and architecture, this adds a build rule that depends on
the `opam-ci-check` executable (so should only update when that changes),
and will embed the commit hash of repo as a value, which we then thread
thru for when we pin the executable.