opam-ci-check: Command to list-revdeps in docker image
The `opam-ci-check list` command can be run locally to list revdeps for
a package. It creates an isolated opam root for some isolation, but for
a greater extent of isolation, this commit adds a list-revdeps command
that can be used to list-revdeps locally using a Docker image.

This makes it easier to reproduce the CI runs more closely, when trying
to reproduce locally. And adds extra flexibility to be able to test on
different OSes etc.