Refactor lint checks to support optional kinds
Extends the Lint api to allow specifying sets of lint checks to run,
allowing us to control which kinds of checks we run in a given case.

The changes in this commit to not add any new checks, but in factoring
out the checks into distinct sets, we do change the order in which lint
errors are reported. The expectations for those tests are updated here
accordingly.