Expose the check kinds from the CLI
Allows specifing subsets of the cli checks to run. Note that this commit
only updates tests to adapt to the new help output on our existing
erroneous CLI input tests. This all the content of tests stays the same.
This helps to show that this CLI change does not break our existing API.