Improve parse errors in linting check
We had an error recently in CI that was tricky to figure out because it
was unclear what kind of error it was (e.g., due to parsing versions or
bad format or something else), and which position it was arising from in
the file.

This improvement reuses `opam-format`'s error reporting logic to give
a more useful errors messages, that includes the position of the error.