Remove double checks for source archive checksums
The opam lint with `--check-upstream` ensures that the checksums for the
upstream archive is correct. But, we also have a separate
`check_checksums` that also ensures this. This commit removes the double
checksum check in the default case, but keeps the check when running
with only "--checks opam-file".

This partially addresses #400.