lower-bounds: run the check with the opam-repository-archive on
Currently the opam-repository will test the lower bound on most latest versions of packages. And even though most cases are avoided by the fact that we add upper bounds to all archived packages, the new packages may still be compatible and co-installable. So we should keep checking the lower bounds with the whole database to ensure that the metadata on the opam-repository remains consistent