Remove dune constraints check from the lint checks
The lint check can produce specious errors, for instance for packages
where the dune-package is not present at the root of the repository. We
remove the check since a problem with the `dune-package` file should be
caught in the build step, and this lint doesn't add anything in the
problematic case, but causes confusion in other cases.

Closes #436